Bug/dropbit invite

See merge request cn/dropbit-android!87

+++ app/src/main/java/com/coinninja/coinkeeper/service/runner/SentInvitesStatusGetter.java
@@ -101,7 +101,9 @@ public class SentInvitesStatusGetter implements Runnable {
 
         for (InviteTransactionSummary unacknowledgedInvitation: allUnacknowledgedInvitations) {
             if(serverIds.get(unacknowledgedInvitation.getId()) == null) {
-                unacknowledgedInvitation.getTransactionsInvitesSummary().delete();
+                if (unacknowledgedInvitation.getTransactionNotification() != null) {
+                    unacknowledgedInvitation.getTransactionsInvitesSummary().delete();
+                }
                 unacknowledgedInvitation.delete();
             }
         }
